converting to static pages
