+++ app/routes/01-lets-start/lesson-01.module.scss
@@ -1,0 +1,286 @@
+@import '../../../src/globals/variables.module.scss';
+
+.root {
+    display: flex;
+    justify-content: center;
+    box-sizing: border-box;
+    background-color: $lesson-bg-color;
+    font-family: $font-family-text;
+    padding: $lesson-padding;
+    gap: $lesson-gap;
+
+    .playground {
+        overflow: hidden;
+        background-color: $playground-bg-color;
+        border-radius: $playground-border-radius;
+        width: $playground-width;
+        height: $playground-height;
+
+        // lesson specific
+        background-image: url('../../../assets/lesson-01-bg.png');
+        background-size: cover;
+        position: relative;
+    }
+}
+
+// lesson specific
+.fadeOut {
+    opacity: 0;
+}
+
+.boxes {
+    position: absolute;
+    border-radius: $box-border-radius-rounded;
+
+    // lavender
+    &.boxes:nth-child(1) {
+        width: 160px;
+        height: 60px;
+        left: 26px;
+        top: 46px;
+
+        z-index: 11;
+        transform: rotate(-37deg);
+    }
+
+    // turquoiseGreen
+    &.boxes:nth-child(2) {
+        width: 180px;
+        height: 60px;
+        left: 28px;
+        top: 48px;
+
+        z-index: 10;
+        transform: rotate(5deg);
+    }
+
+    // dirtyWhite
+    &.boxes:nth-child(3) {
+        width: 100px;
+        height: 60px;
+        left: 164px;
+        top: 70px;
+
+        z-index: 10;
+        transform: rotate(-32deg);
+    }
+
+    // aeroBlue
+    &.boxes:nth-child(4) {
+        width: 60px;
+        height: 60px;
+        left: 242px;
+        top: 62px;
+
+        z-index: 10;
+        border-radius: $box-border-radius-circle;
+    }
+
+    // redStop
+    &.boxes:nth-child(5) {
+        width: 65px;
+        height: 65px;
+        left: 78px;
+        top: 94px;
+
+        z-index: 13;
+        border-radius: $box-border-radius-circle;
+    }
+
+    // hotPink
+    &.boxes:nth-child(6) {
+        width: 60px;
+        height: 60px;
+        left: 162px;
+        top: 82px;
+
+        z-index: 11;
+        border-radius: $box-border-radius-circle;
+    }
+
+    // garudaGold
+    &.boxes:nth-child(7) {
+        width: 100px;
+        height: 100px;
+        left: 108px;
+        top: 120px;
+
+        z-index: 12;
+        border-radius: $box-border-radius-circle;
+    }
+
+    // orangeRed
+    &.boxes:nth-child(8) {
+        width: 110px;
+        height: 60px;
+        left: 200px;
+        top: 136px;
+
+        z-index: 11;
+    }
+
+    // aeroBlue
+    &.boxes:nth-child(9) {
+        width: 180px;
+        height: 60px;
+        left: 36px;
+        top: 148px;
+
+        z-index: 10;
+        transform: rotate(-18deg);
+    }
+
+    // richGreen
+    &.boxes:nth-child(10) {
+        width: 85px;
+        height: 85px;
+        left: 60px;
+        top: 190px;
+
+        z-index: 10;
+        border-radius: $box-border-radius-circle;
+    }
+
+    // darkLavender
+    &.boxes:nth-child(11) {
+        width: 120px;
+        height: 60px;
+        left: 246px;
+        top: 174px;
+
+        z-index: 10;
+        transform: rotate(-14deg);
+    }
+
+    // oliveGreen
+    &.boxes:nth-child(12) {
+        width: 82px;
+        height: 82px;
+        left: 174px;
+        top: 220px;
+
+        z-index: 13;
+        border-radius: $box-border-radius-circle;
+    }
+
+    // orangeRed
+    &.boxes:nth-child(13) {
+        width: 190px;
+        height: 60px;
+        left: 22px;
+        top: 276px;
+
+        z-index: 13;
+        transform: rotate(16deg);
+    }
+
+    // hotPink
+    &.boxes:nth-child(14) {
+        width: 206px;
+        height: 60px;
+        left: 50px;
+        top: 278px;
+
+        z-index: 12;
+        transform: rotate(-6deg);
+    }
+
+    // dirtyWhite
+    &.boxes:nth-child(15) {
+        width: 266px;
+        height: 60px;
+        left: 126px;
+        top: 302px;
+
+        z-index: 11;
+        transform: rotate(-22deg);
+    }
+
+    // richGreen
+    &.boxes:nth-child(16) {
+        width: 150px;
+        height: 60px;
+        left: 248px;
+        top: 346px;
+
+        z-index: 13;
+        transform: rotate(52deg);
+    }
+
+    // oliveGreen
+    &.boxes:nth-child(17) {
+        width: 180px;
+        height: 60px;
+        left: 28px;
+        top: 366px;
+
+        z-index: 10;
+        transform: rotate(12deg);
+    }
+
+    // orangeRed
+    &.boxes:nth-child(18) {
+        width: 60px;
+        height: 60px;
+        left: 230px;
+        top: 357px;
+
+        z-index: 12;
+        border-radius: $box-border-radius-circle;
+    }
+
+    // turquoiseGreen
+    &.boxes:nth-child(19) {
+        width: 146px;
+        height: 60px;
+        left: 172px;
+        top: 394px;
+
+        z-index: 11;
+        transform: rotate(18deg);
+    }
+
+    // aeroBlue
+    &.boxes:nth-child(20) {
+        width: 80px;
+        height: 80px;
+        left: 56px;
+        top: 406px;
+
+        z-index: 11;
+        border-radius: $box-border-radius-circle;
+    }
+
+    // darkLavender
+    &.boxes:nth-child(21) {
+        width: 80px;
+        height: 60px;
+        left: 126px;
+        top: 414px;
+
+        z-index: 13;
+        transform: rotate(11deg);
+    }
+
+    // redStop
+    &.boxes:nth-child(22) {
+        width: 60px;
+        height: 60px;
+        left: 160px;
+        top: 410px;
+
+        z-index: 12;
+        border-radius: $box-border-radius-circle;
+    }
+
+    // hotPink
+    &.boxes:nth-child(23) {
+        width: 176px;
+        height: 60px;
+        left: 100px;
+        top: 438px;
+
+        z-index: 11;
+        transform: rotate(-14deg);
+    }
+}
