+++ app/routes/03-add-elements/lesson-03.module.scss
@@ -1,0 +1,41 @@
+@import '../../../src/globals/variables.module.scss';
+
+.root {
+    display: flex;
+    justify-content: center;
+    box-sizing: border-box;
+    background-color: $lesson-bg-color;
+    font-family: $font-family-text;
+    padding: $lesson-padding;
+    gap: $lesson-gap;
+
+    .playground {
+        overflow: hidden;
+        border-radius: $playground-border-radius;
+        width: $playground-width;
+        height: $playground-height;
+
+        // lesson specific
+        position: relative;
+        display: flex;
+        flex-direction: column;
+    }
+}
+
+// lesson specific
+.eye {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 280px;
+
+    &.top {
+        background-image: url('../../../assets/eye.svg');
+        background-color: $charcoal-black;
+    }
+
+    &.bottom {
+        background-image: url('../../../assets/eye.svg');
+        background-color: $gold-yellow;
+    }
+}
