+++ app/routes/03-add-elements/pupil/pupil.module.scss
@@ -1,0 +1,29 @@
+@import '../../../../src/globals/variables.module.scss';
+
+.root {
+  font-family: $font-family-text;
+}
+
+.pupils {
+  box-sizing: border-box;
+  align-items: center;
+  justify-items: center;
+  display: grid;
+  grid-template-columns: 70px;
+  grid-template-rows: 70px;
+}
+
+.pupil {
+  border-radius: $box-border-radius-circle;
+  grid-column: 1/1;
+  grid-row: 1/1;
+
+  &.pupil:nth-child(1) {
+    width: 70px;
+    height: 70px;
+  }
+  &.pupil:nth-child(2) {
+    width: 46px;
+    height: 46px;
+  }
+}
