+++ app/routes/05-styles-panel/lesson-05.module.scss
@@ -1,0 +1,59 @@
+@import '../../../src/globals/variables.module.scss';
+
+.root {
+    display: flex;
+    justify-content: center;
+    box-sizing: border-box;
+    background-color: $lesson-bg-color;
+    font-family: $font-family-text;
+    padding: $lesson-padding;
+    gap: $lesson-gap;
+
+    .playground {
+        overflow: hidden;
+        background-color: $snow-white;
+        border-radius: $playground-border-radius;
+        width: $playground-width;
+        height: $playground-height;
+
+        // lesson specific
+        position: relative;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: space-between;
+    }
+}
+
+// lesson specific
+.section {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-color: $charcoal-black;
+    border-radius: $playground-border-radius;
+    width: 400px;
+    height: 276px;
+
+    .row {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+
+        .margin {
+            border-radius: $box-border-radius-circle;
+            font-style: italic;
+            font-weight: 800;
+            font-size: 35px;
+            width: 70px;
+            height: 70px;
+            margin: 35px;
+        }
+
+        .padding {
+            border-radius: $box-border-radius-rounded;
+            font-style: italic;
+            font-size: 20px;
+        }
+    }
+}
