+++ app/routes/06-create-class/lesson-06.module.scss
@@ -1,0 +1,45 @@
+@import '../../../src/globals/variables.module.scss';
+
+.root {
+    display: flex;
+    justify-content: center;
+    box-sizing: border-box;
+    background-color: $lesson-bg-color;
+    font-family: $font-family-text;
+    padding: $lesson-padding;
+    gap: $lesson-gap;
+}
+
+.playground {
+    overflow: hidden;
+    border-radius: $playground-border-radius;
+    width: $playground-width;
+    height: $playground-height;
+
+    // lesson specific
+    position: relative;
+    box-sizing: border-box;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    padding: 128px 75px 92px 75px;
+
+    font-style: italic;
+    font-weight: 500;
+    font-size: 45px;
+    line-height: 55px;
+    color: $orange-red;
+    background-color: $pastel-pink;
+}
+
+// lesson specific
+.quote {
+    span {
+        color: $pastel-pink;
+    }
+}
+
+.author {
+    color: $pastel-pink;
+    font-size: 18px;
+}
