+++ app/routes/07-components/lesson-07.module.scss
@@ -1,0 +1,34 @@
+@import '../../../src/globals/variables.module.scss';
+
+.root {
+    display: flex;
+    justify-content: center;
+    box-sizing: border-box;
+    background-color: $lesson-bg-color;
+    font-family: $font-family-text;
+    padding: $lesson-padding;
+    gap: $lesson-gap;
+
+    .playground {
+        overflow: hidden;
+        background-color: $playground-bg-color;
+        border-radius: $playground-border-radius;
+        width: $playground-width;
+        height: $playground-height;
+
+        // lesson specific
+        position: relative;
+        box-sizing: border-box;
+        display: grid;
+        grid-template-columns: 60px 60px 60px;
+        grid-template-rows: 60px 60px 60px;
+        padding: 160px 80px 160px 80px;
+        gap: 30px;
+    }
+}
+
+// lesson specific
+.icon {
+    border-radius: $box-border-radius-circle;
+    padding: 12px;
+}
