+++ app/routes/08-flex-layouts/lesson-08.module.scss
@@ -1,0 +1,61 @@
+@import '../../../src/globals/variables.module.scss';
+
+.root {
+    display: flex;
+    justify-content: center;
+    box-sizing: border-box;
+    background-color: $lesson-bg-color;
+    font-family: $font-family-text;
+    padding: $lesson-padding;
+    gap: $lesson-gap;
+
+    .playground {
+        overflow: hidden;
+        background-color: $playground-bg-color;
+        border-radius: $playground-border-radius;
+        width: $playground-width;
+        height: $playground-height;
+
+        // lesson specific
+        position: relative;
+    }
+}
+
+// lesson specific
+.section {
+    display: flex;
+    box-sizing: border-box;
+    position: absolute;
+    width: 400px;
+    height: 560px;
+}
+
+.tags {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    gap: 10px;
+    padding: 55px 50px 50px 50px;
+
+    .tag {
+        border-radius: $box-border-radius-circle;
+        padding: 2px;
+        height: 80px;
+    }
+}
+
+// solution
+//flex-direction: column;, justify-content: flex-end, gap: 26px
+.circles {
+    display: flex;
+    justify-content: space-between;
+    padding: 55px 50px 58px 58px;
+    gap: 26px;
+    z-index: 1;
+
+    .circle {
+        border-radius: $box-border-radius-circle;
+        width: 64px;
+        height: 64px;
+    }
+}
