+++ codux.config.json
@@ -2,14 +2,8 @@
   "$schema": "https://wixplosives.github.io/codux-config-schema/codux.config.schema.json",
   "boardGlobalSetup": "./_codux/boards-global-setup.ts",
   "componentsDiscovery": {
-    "include": ["src/components/**"],
-    "exclude": [
-      "src/_codux/component-templates/**",
-      "src/components/common/**",
-      "src/components/tasks/**",
-      "src/components/fx/**",
-      "src/App.tsx"
-    ]
+    "include": ["src/**"],
+    "exclude": ["src/_codux/component-templates/**"]
   },
   "newComponent": {
     "componentsPath": "src/components",
