--- src/components/lessons/02/lesson-02.module.scss
@@ -1,76 +1,0 @@
-@import '../../../globals/variables.module.scss';
-
-.root {
-  display: flex;
-  justify-content: center;
-  box-sizing: border-box;
-  background-color: $lesson-bg-color;
-  font-family: $font-family-text;
-  padding: $lesson-padding;
-  gap: $lesson-gap;
-
-  .playground {
-    overflow: hidden;
-    background-color: $playground-bg-color;
-    border-radius: $playground-border-radius;
-    width: $playground-width;
-    height: $playground-height;
-
-    // lesson specific
-    position: relative;
-    box-sizing: border-box;
-    display: grid;
-    grid-template-columns: 250px;
-    grid-template-rows: 60px 60px 60px 60px;
-    padding: 130px 75px 130px 75px;
-    gap: 20px;
-  }
-}
-
-// lesson specific
-.row {
-  display: flex;
-  justify-content: space-between;
-  color: $charcoal-black;
-  font-size: 1.875rem;
-
-  .box01 {
-    border-radius: $box-border-radius-circle;
-    width: 60px;
-  }
-
-  .box02 {
-    border-radius: $box-border-radius-rounded;
-    width: 180px;
-  }
-
-  .box03 {
-    border-radius: $box-border-radius-rounded;
-    width: 120px;
-  }
-
-  .box04 {
-    border-radius: $box-border-radius-rounded;
-    width: 120px;
-  }
-
-  .box05 {
-    border-radius: $box-border-radius-rounded;
-    width: 250px;
-  }
-
-  .box06 {
-    border-radius: $box-border-radius-circle;
-    width: 60px;
-  }
-
-  .box07 {
-    border-radius: $box-border-radius-rounded;
-    width: 110px;
-  }
-
-  .box08 {
-    border-radius: $box-border-radius-circle;
-    width: 60px;
-  }
-}
