--- src/components/lessons/04/lesson-04.module.scss
@@ -1,74 +1,0 @@
-@import '../../../globals/variables.module.scss';
-
-.root {
-  display: flex;
-  justify-content: center;
-  box-sizing: border-box;
-  background-color: $lesson-bg-color;
-  font-family: $font-family-text;
-  padding: $lesson-padding;
-  gap: $lesson-gap;
-
-  .playground {
-    overflow: hidden;
-    background-color: $playground-bg-color;
-    border-radius: $playground-border-radius;
-    width: $playground-width;
-    height: $playground-height;
-
-    // lesson specific
-    position: relative;
-    box-sizing: border-box;
-    display: grid;
-    grid-template-columns: auto auto;
-    padding: 40px;
-    gap: 1px;
-  }
-}
-
-// lesson specific
-.column {
-  width: 160px;
-  height: 480px;
-}
-
-.images {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 1px;
-}
-
-.left {
-  flex: 1 1 auto;
-  width: 160px;
-
-  &.left:nth-child(1) {
-    height: 16px;
-  }
-
-  &.left:nth-child(2) {
-    height: 120px;
-  }
-
-  &.left:nth-child(3) {
-    height: 240px;
-  }
-}
-
-.right {
-  flex: 1 1 auto;
-  width: 160px;
-
-  &.right:nth-child(1) {
-    height: 10px;
-  }
-
-  &.right:nth-child(2) {
-    height: 200px;
-  }
-
-  &.right:nth-child(3) {
-    height: 100px;
-  }
-}
