--- src/components/lessons/10/lesson-10.module.scss
@@ -1,69 +1,0 @@
-@import '../../../globals/variables.module.scss';
-
-.root {
-  display: flex;
-  justify-content: center;
-  box-sizing: border-box;
-  background-color: $lesson-bg-color;
-  font-family: $font-family-text;
-  padding: $lesson-padding;
-  gap: $lesson-gap;
-
-  .playground {
-    overflow: hidden;
-    // workaround to fix background bleed (when border-radius)
-    background: linear-gradient(
-      $playground-bg-color,
-      $playground-bg-color 95%,
-      $lesson-bg-color 95%
-    );
-    border-radius: $playground-border-radius;
-    width: $playground-width;
-    height: $playground-height;
-
-    // lesson specific
-    position: relative;
-  }
-}
-
-// lesson specific
-.arcs {
-  position: absolute;
-  width: 702px;
-  height: 702px;
-  left: 50px;
-  top: 50px;
-  background: radial-gradient(
-    $charcoal-black 80px,
-    $olive-green 80px,
-    $olive-green 110px,
-    $dirty-white 110px,
-    $dirty-white 140px,
-    $turquoise-green 140px,
-    $turquoise-green 170px,
-    $rich-green 170px,
-    $rich-green 200px,
-    $aero-blue 200px,
-    $aero-blue 230px,
-    $dark-lavender 230px,
-    $dark-lavender 260px,
-    $hot-pink 260px,
-    $hot-pink 290px,
-    $pastel-pink 290px,
-    $pastel-pink 320px,
-    $hot-red 320px,
-    $hot-red 350px,
-    rgba(0, 0, 0, 0) 350px,
-    rgba(0, 0, 0, 0)
-  );
-}
-
-.cloud {
-  position: absolute;
-  fill: $dana-pink;
-  width: $playground-width;
-  height: 260px;
-  left: 0px;
-  top: 300px;
-  z-index: 1;
-}
