--- src/components/lessons/10/lesson-10.tsx
@@ -1,15 +1,0 @@
-import Cloud from '../../../assets/cloud.svg?react';
-import { Task10 as Task } from '../../tasks/10/task-10';
-import styles from './lesson-10.module.scss';
-
-export const Lesson10 = () => {
-    return (
-        <div className={styles.root}>
-            <Task />
-            <div className={styles.playground}>
-                <div className={styles.arcs} />
-                <Cloud className={styles.cloud} />
-            </div>
-        </div>
-    );
-};
