--- src/components/lessons/default-page/default-page.tsx
@@ -1,8 +1,0 @@
-export const DefaultPage = () => {
-    return (
-        <div>
-            <h1>Default Page</h1>
-            <p>This is the default page for the lessons.</p>
-        </div>
-    );
-};
