--- src/components/lessons/index.ts
@@ -1,11 +1,0 @@
-export { Lesson01 } from './01/lesson-01';
-export { Lesson02 } from './02/lesson-02';
-export { Lesson03 } from './03/lesson-03';
-export { Lesson04 } from './04/lesson-04';
-export { Lesson05 } from './05/lesson-05';
-export { Lesson06 } from './06/lesson-06';
-export { Lesson07 } from './07/lesson-07';
-export { Lesson08 } from './08/lesson-08';
-export { Lesson09 } from './09/lesson-09';
-export { Lesson10 } from './10/lesson-10';
-export { DefaultPage } from './default-page/default-page';
